refactor: Accommodation 엔티티 패키지 변경, Reservation 연관관계 삭제 (ios-h/airbnb#19)
