+++ BE/src/main/java/org/team4/airbnb/domain/AccommodationImage.java
@@ -1,6 +1,14 @@
 package org.team4.airbnb.domain;
 
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import org.team4.airbnb.accommodation.Accommodation;
 
 @Entity
 public class AccommodationImage {
