+++ BE/src/main/java/org/team4/airbnb/domain/Reservation.java
@@ -1,7 +1,16 @@
 package org.team4.airbnb.domain;
 
-import javax.persistence.*;
 import java.time.LocalDate;
+import javax.persistence.Column;
+import javax.persistence.Embedded;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import org.team4.airbnb.accommodation.Accommodation;
 import org.team4.airbnb.customer.Customer;
 
 @Entity
